+++ rlap/cc/samplers.h
@@ -9,21 +9,21 @@
 class CliqueSampler{
     // Samplers to approximate the CLIQUE induced by removing nodes in the graph
   public:
-    virtual Eigen::MatrixXf sampleClique(Eigen::SparseMatrix<float>& L, int k) = 0;
+    virtual Eigen::MatrixXd sampleClique(Eigen::SparseMatrix<double>& L, int k) = 0;
 };
 
 class ExactCliqueSampler: public CliqueSampler{
   public:
     ExactCliqueSampler(){};
     ~ExactCliqueSampler(){};
-    Eigen::MatrixXf sampleClique(Eigen::SparseMatrix<float>& L, int i) override;
+    Eigen::MatrixXd sampleClique(Eigen::SparseMatrix<double>& L, int i) override;
 };
 
 class WeightedCliqueSampler: public CliqueSampler{
   public:
     WeightedCliqueSampler(){};
     ~WeightedCliqueSampler(){};
-    Eigen::MatrixXf sampleClique(Eigen::SparseMatrix<float>& L, int i) override;
+    Eigen::MatrixXd sampleClique(Eigen::SparseMatrix<double>& L, int i) override;
 };
